Rename "jira_updater" to "updater_for_jira".
Change the color of our bike shed to something slightly different.
