+++ updater_for_jira/jirahelper.py
