+++ updater_for_jira/views.py
